% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{are_orthogonals}
\alias{are_orthogonals}
\title{Check if two columns of a dataframe are orthogonal}
\usage{
are_orthogonals(data, C1, C2)
}
\arguments{
\item{data}{dataframe}

\item{C1}{column name}

\item{C2}{column name}
}
\value{
boolean
}
\description{
Check if two columns of a dataframe are orthogonal
}
