% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{CTCRW_cov}
\alias{CTCRW_cov}
\title{Covariance matrix of CTCRW transition density}
\usage{
CTCRW_cov(beta, sigma, dt)
}
\arguments{
\item{beta}{Parameter beta of CTCRW model (mean reversion)}

\item{sigma}{Parameter sigma of CTCRW model (diffusion)}

\item{dt}{Time interval}
}
\value{
Covariance matrix of position/velocity joint 
process
}
\description{
The derivation of these equations is for example included in
Michelot et al. (2019), Stochastic models of animal movement and
habitat selection (Sections 6.2.2.3-6.2.2.4).
}
