% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{edf}
\alias{edf}
\title{EDF for smooth terms = trace(F)}
\usage{
edf(X_re, S, lambda, ncol_re)
}
\arguments{
\item{X_re}{Design matrix for random effects}

\item{S}{Smoothing matrix}

\item{lambda}{Smoothing parameter}

\item{ncol_re}{Number of columns of S for each random effect}
}
\value{
Trace of F = (Xt X + sp*S)^-1 Xt X
}
\description{
This function is adapted from Dave Miller's code in the
package CTMCdive
}
