% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{edf_smooth}
\alias{edf_smooth}
\title{Compute effective degrees of freedom for a smooth term (i.e., GAM)}
\usage{
edf_smooth(X_re, S, lambda)
}
\arguments{
\item{X_re}{Design matrix for random effects}

\item{S}{Smoothing matrix}

\item{lambda}{Smoothing parameter}
}
\value{
Trace of F = (Xt X + sp*S)^-1 Xt X
}
\description{
This function is adapted from Dave Miller's code in the
package CTMCdive
}
