% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_BoundaryMetrics}
\alias{get_BoundaryMetrics}
\title{Compute distance to boundary and angle for CRCVM_SSM sde type}
\usage{
get_BoundaryMetrics(data, response, border, n_cores = NULL)
}
\arguments{
\item{data}{dataframe with columns time, ID and response for response variable (such as animal positions)}

\item{response}{name of response variables}

\item{border}{sf object with polygons geometry defining zones that the process cannot reach.}

\item{n_cores}{number of cores to use for parallelization}
}
\value{
dataframe with the same number of rows as data, and two columns BoundaryDistance and BoundaryAngle.
}
\description{
Compute distance to boundary and angle for CRCVM_SSM sde type
}
