% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_variables}
\alias{get_variables}
\title{Get covariable names from a formula}
\usage{
get_variables(formula)
}
\arguments{
\item{formula}{a formula (usually for one parameter of the SDE)}
}
\value{
Vector of names of the covariables in the formula
}
\description{
Get covariable names from a formula
}
