% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{interpolate_BoundaryMetrics}
\alias{interpolate_BoundaryMetrics}
\title{Interpolate BoundaryDistance and BoundaryAngle for CRCVM (need response with dimension 2)}
\usage{
interpolate_BoundaryMetrics(
  data,
  response,
  border,
  n_step = 1,
  sp = NULL,
  df_ratio = NULL,
  n_cores = NULL
)
}
\arguments{
\item{data}{Dataframe with the columns time, ID, BoundaryDistance, and BoundaryAngle}

\item{response}{name of response variables}

\item{border}{Boundary of the domain as a sf object with geometry}

\item{n_step}{Number of intervals to subdivise between two observations}

\item{sp}{smoothing penalty for the smoothing of the trajectories}

\item{df_ratio}{ratio of degree of freedom over number of data points to use, must be <1}

\item{n_cores}{number of cores to use for computation in parallel mode}
}
\value{
List with matrices of size nrow(data) x n_step that contain interpolated positions, distances
angles and interpolation times
}
\description{
Interpolate BoundaryDistance and BoundaryAngle for CRCVM (need response with dimension 2)
}
