% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is_in_border}
\alias{is_in_border}
\title{Check whether a point is on border or not}
\usage{
is_in_border(point, border)
}
\arguments{
\item{point}{the location we want to tell if it is on border or not}

\item{border}{sf object (list of polygons) defining the border}
}
\value{
boolean TRUE if it is on border, FALSE otherwise
}
\description{
Check whether a point is on border or not
}
