% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is_in_land}
\alias{is_in_land}
\title{Check whether a point is on land or not}
\usage{
is_in_land(point, land)
}
\arguments{
\item{point}{the location we want to tell if it is on land or not}

\item{land}{sf object (list of polygons) defining the land}
}
\value{
boolean TRUE if it is on land, FALSE otherwise
}
\description{
Check whether a point is on land or not
}
