% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{logLik.SDE}
\alias{logLik.SDE}
\title{logLik function for SDE objects}
\usage{
\method{logLik}{SDE}(object, ...)
}
\arguments{
\item{object}{SDE model object}
}
\value{
Maximum log-likelihood value for the model, with attributes
\code{df} (degrees of freedom) and \code{nobs} (number of observations)
}
\description{
This function makes it possible to call generic R methods such
as AIC and BIC on SDE objects.
}
