% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{nearest_boundary_point}
\alias{nearest_boundary_point}
\title{Compute nearest point on the shoreline}
\usage{
nearest_boundary_point(point, land)
}
\arguments{
\item{point}{the location from which we want to find the nearest point on land}

\item{land}{sf object (list of polygons) defining the land}
}
\value{
matrix with on row and two columns that are the coordinates of the nearest point in land
}
\description{
Compute nearest point on the shoreline
}
