% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{nearest_boundary_points}
\alias{nearest_boundary_points}
\title{Compute nearest points on the shoreline}
\usage{
nearest_boundary_points(points, border)
}
\arguments{
\item{border}{sf object (list of polygons) defining the border}

\item{point}{the location from which we want to find the nearest point on border}
}
\value{
matrix with on row and two columns that are the coordinates of the nearest point in border
}
\description{
Compute nearest points on the shoreline
}
