% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{RACVM_cov}
\alias{RACVM_cov}
\title{Covariance matrix of RACVM transition density}
\usage{
RACVM_cov(beta, sigma, omega, dt)
}
\arguments{
\item{beta}{Parameter beta of RACVM model (mean reversion)}

\item{sigma}{Parameter sigma of RACVM model (diffusion)}

\item{omega}{Parameter omega of RACVM (angular speed)}

\item{dt}{Time interval}
}
\value{
Covariance matrix of position/velocity joint 
process
}
\description{
Covariance matrix of RACVM transition density
}
