% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{RACVM_drift}
\alias{RACVM_drift}
\title{Drift matrix of RACVM transition density}
\usage{
RACVM_drift(beta, omega, dt)
}
\arguments{
\item{beta}{Parameter beta of RACVM model (mean reversion)}

\item{omega}{Parameter omega of RACVM (angular speed)}

\item{dt}{Time interval}

\item{sigma}{Parameter sigma of RACVM model (diffusion)}
}
\value{
Link matrix for drift term of position/velocity joint 
process
}
\description{
Drift matrix of RACVM transition density
}
