% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{RACVM_link}
\alias{RACVM_link}
\title{Link matrix of RACVM transition density}
\usage{
RACVM_link(beta, omega, dt)
}
\arguments{
\item{beta}{Parameter beta of RACVM model (mean reversion)}

\item{omega}{Parameter omega of RACVM (angular speed)}

\item{dt}{Time interval}

\item{sigma}{Parameter sigma of RACVM model (diffusion)}
}
\value{
Link matrix of position/velocity joint 
process
}
\description{
Link matrix of RACVM transition density
}
