% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sde.R
\name{SDE}
\alias{SDE}
\title{R6 class for stochastic differential equation}
\description{
R6 class for stochastic differential equation

R6 class for stochastic differential equation
}
\details{
Contains the model formulas and data.

This function transforms the smoothness parameter of
each smooth term into a standard deviation, given by 
SD = 1/sqrt(lambda). It is particularly helpful to get the
standard deviations of independent normal random effects.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{SDE$new()}}
\item \href{#method-formulas}{\code{SDE$formulas()}}
\item \href{#method-data}{\code{SDE$data()}}
\item \href{#method-type}{\code{SDE$type()}}
\item \href{#method-response}{\code{SDE$response()}}
\item \href{#method-fixpar}{\code{SDE$fixpar()}}
\item \href{#method-link}{\code{SDE$link()}}
\item \href{#method-invlink}{\code{SDE$invlink()}}
\item \href{#method-coeff_fe}{\code{SDE$coeff_fe()}}
\item \href{#method-coeff_re}{\code{SDE$coeff_re()}}
\item \href{#method-lambda}{\code{SDE$lambda()}}
\item \href{#method-vcomp}{\code{SDE$vcomp()}}
\item \href{#method-terms}{\code{SDE$terms()}}
\item \href{#method-res}{\code{SDE$res()}}
\item \href{#method-tmb_obj}{\code{SDE$tmb_obj()}}
\item \href{#method-tmb_rep}{\code{SDE$tmb_rep()}}
\item \href{#method-obs}{\code{SDE$obs()}}
\item \href{#method-eqn}{\code{SDE$eqn()}}
\item \href{#method-update_coeff_fe}{\code{SDE$update_coeff_fe()}}
\item \href{#method-update_coeff_re}{\code{SDE$update_coeff_re()}}
\item \href{#method-update_lambda}{\code{SDE$update_lambda()}}
\item \href{#method-message}{\code{SDE$message()}}
\item \href{#method-ind_fixcoeff}{\code{SDE$ind_fixcoeff()}}
\item \href{#method-make_mat}{\code{SDE$make_mat()}}
\item \href{#method-make_mat_grid}{\code{SDE$make_mat_grid()}}
\item \href{#method-setup}{\code{SDE$setup()}}
\item \href{#method-fit}{\code{SDE$fit()}}
\item \href{#method-par_all}{\code{SDE$par_all()}}
\item \href{#method-post}{\code{SDE$post()}}
\item \href{#method-residuals}{\code{SDE$residuals()}}
\item \href{#method-AIC}{\code{SDE$AIC()}}
\item \href{#method-plot_par}{\code{SDE$plot_par()}}
\item \href{#method-clone}{\code{SDE$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a SDE object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$new(formulas, data, type, response, par0 = NULL, fixpar = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{formulas}}{List of formulas for model parameters}

\item{\code{data}}{Data frame with covariates, response variable,
time, and ID}

\item{\code{type}}{Type of SDE ("BM", "OU"...)}

\item{\code{response}}{Vector of names of response variables}

\item{\code{par0}}{Vector of initial values for SDE parameters}

\item{\code{fixpar}}{Vector of names of fixed SDE parameters}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new SDE object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-formulas"></a>}}
\if{latex}{\out{\hypertarget{method-formulas}{}}}
\subsection{Method \code{formulas()}}{
Formulas of SDE object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$formulas()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data"></a>}}
\if{latex}{\out{\hypertarget{method-data}{}}}
\subsection{Method \code{data()}}{
Data of SDE object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-type"></a>}}
\if{latex}{\out{\hypertarget{method-type}{}}}
\subsection{Method \code{type()}}{
Type of SDE object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-response"></a>}}
\if{latex}{\out{\hypertarget{method-response}{}}}
\subsection{Method \code{response()}}{
Name(s) of response variable(s)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$response()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fixpar"></a>}}
\if{latex}{\out{\hypertarget{method-fixpar}{}}}
\subsection{Method \code{fixpar()}}{
Name(s) of fixed parameter(s)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$fixpar()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-link"></a>}}
\if{latex}{\out{\hypertarget{method-link}{}}}
\subsection{Method \code{link()}}{
Link functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$link()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-invlink"></a>}}
\if{latex}{\out{\hypertarget{method-invlink}{}}}
\subsection{Method \code{invlink()}}{
Inverse link functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$invlink()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coeff_fe"></a>}}
\if{latex}{\out{\hypertarget{method-coeff_fe}{}}}
\subsection{Method \code{coeff_fe()}}{
Fixed effect parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$coeff_fe()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-coeff_re"></a>}}
\if{latex}{\out{\hypertarget{method-coeff_re}{}}}
\subsection{Method \code{coeff_re()}}{
Random effect parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$coeff_re()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-lambda"></a>}}
\if{latex}{\out{\hypertarget{method-lambda}{}}}
\subsection{Method \code{lambda()}}{
Smoothness parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$lambda()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-vcomp"></a>}}
\if{latex}{\out{\hypertarget{method-vcomp}{}}}
\subsection{Method \code{vcomp()}}{
Variance components of smooth terms
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$vcomp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-terms"></a>}}
\if{latex}{\out{\hypertarget{method-terms}{}}}
\subsection{Method \code{terms()}}{
Terms of model formulas
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$terms()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-res"></a>}}
\if{latex}{\out{\hypertarget{method-res}{}}}
\subsection{Method \code{res()}}{
Output of optimiser after model fitting
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$res()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmb_obj"></a>}}
\if{latex}{\out{\hypertarget{method-tmb_obj}{}}}
\subsection{Method \code{tmb_obj()}}{
Model object created by TMB. This is the output of 
the TMB function \code{MakeADFun}, and it is a list including elements
\itemize{
  \item{\code{fn}}{Objective function}
  \item{\code{gr}}{Gradient function of fn}
  \item{\code{par}}{Vector of initial parameters on working scale}
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$tmb_obj()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-tmb_rep"></a>}}
\if{latex}{\out{\hypertarget{method-tmb_rep}{}}}
\subsection{Method \code{tmb_rep()}}{
Output of the TMB function \code{sdreport}, which includes 
estimates and standard errors for all model parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$tmb_rep()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-obs"></a>}}
\if{latex}{\out{\hypertarget{method-obs}{}}}
\subsection{Method \code{obs()}}{
Data frame of observations (subset response
variables out of full data frame)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$obs()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eqn"></a>}}
\if{latex}{\out{\hypertarget{method-eqn}{}}}
\subsection{Method \code{eqn()}}{
Print equation for this model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$eqn()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_coeff_fe"></a>}}
\if{latex}{\out{\hypertarget{method-update_coeff_fe}{}}}
\subsection{Method \code{update_coeff_fe()}}{
Update fixed effect coefficients
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$update_coeff_fe(new_coeff)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_coeff}}{New coefficient vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_coeff_re"></a>}}
\if{latex}{\out{\hypertarget{method-update_coeff_re}{}}}
\subsection{Method \code{update_coeff_re()}}{
Update random effect coefficients
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$update_coeff_re(new_coeff)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_coeff}}{New coefficient vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_lambda"></a>}}
\if{latex}{\out{\hypertarget{method-update_lambda}{}}}
\subsection{Method \code{update_lambda()}}{
Update smoothness parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$update_lambda(new_lambda)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_coeff}}{New smoothness parameter vector}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-message"></a>}}
\if{latex}{\out{\hypertarget{method-message}{}}}
\subsection{Method \code{message()}}{
Print SDE and parameter formulas
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$message()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ind_fixcoeff"></a>}}
\if{latex}{\out{\hypertarget{method-ind_fixcoeff}{}}}
\subsection{Method \code{ind_fixcoeff()}}{
Indices of fixed coefficients in coff_fe
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$ind_fixcoeff()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_mat"></a>}}
\if{latex}{\out{\hypertarget{method-make_mat}{}}}
\subsection{Method \code{make_mat()}}{
Create model matrices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$make_mat(new_data = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{new_data}}{Optional new data set, including covariates for which
the design matrices should be created.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of
\itemize{
  \item X_fe Design matrix for fixed effects
  \item X_re Design matrix for random effects
  \item S Smoothness matrix
  \item ncol_fe Number of columns for X_fe for each parameter
  \item ncol_re Number of columns of X_re and S for each random effect
}
Design matrices for grid of covariates
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-make_mat_grid"></a>}}
\if{latex}{\out{\hypertarget{method-make_mat_grid}{}}}
\subsection{Method \code{make_mat_grid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$make_mat_grid(var, covs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{Name of variable}

\item{\code{covs}}{Optional data frame with a single row and one column
for each covariate, giving the values that should be used. If this is
not specified, the mean value is used for numeric variables, and the
first level for factor variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with the same elements as the output of make_mat, 
plus a data frame of covariates values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-setup"></a>}}
\if{latex}{\out{\hypertarget{method-setup}{}}}
\subsection{Method \code{setup()}}{
TMB setup
 
This creates an attribute \code{tmb_obj}, which can be used to 
evaluate the negative log-likelihood function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$setup(silent = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{Logical. If TRUE, all tracing outputs are hidden (default).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fit"></a>}}
\if{latex}{\out{\hypertarget{method-fit}{}}}
\subsection{Method \code{fit()}}{
Model fitting

The negative log-likelihood of the model is minimised using the
function \code{optim}. TMB uses the Laplace approximation to integrate 
the random effects out of the likelihood.

After the model has been fitted, the output of \code{optim} can be
accessed using the method \code{res}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$fit(silent = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{silent}}{Logical. If TRUE, all tracing outputs are hidden (default).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-par_all"></a>}}
\if{latex}{\out{\hypertarget{method-par_all}{}}}
\subsection{Method \code{par_all()}}{
Get parameters from design matrices
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$par_all(X_fe, X_re)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_fe}}{Design matrix for fixed effects, as returned
by \code{make_mat}}

\item{\code{X_re}}{Design matrix for random effects, as returned
by \code{make_mat}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Matrix with one column for each parameter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post"></a>}}
\if{latex}{\out{\hypertarget{method-post}{}}}
\subsection{Method \code{post()}}{
Posterior draws for uncertainty quantification
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$post(X_fe, X_re, n_post = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_fe}}{Design matrix (fixed effects)}

\item{\code{X_re}}{Design matrix (random effects)}

\item{\code{n_post}}{Number of posterior draws}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-residuals"></a>}}
\if{latex}{\out{\hypertarget{method-residuals}{}}}
\subsection{Method \code{residuals()}}{
Model residuals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$residuals()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AIC"></a>}}
\if{latex}{\out{\hypertarget{method-AIC}{}}}
\subsection{Method \code{AIC()}}{
Akaike Information Criterion

This function is adapted from Dave Miller's code in the
package CTMCdive
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$AIC(k = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{Penalty per parameter; the default 
\code{k = 2} is the classical AIC}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_par"></a>}}
\if{latex}{\out{\hypertarget{method-plot_par}{}}}
\subsection{Method \code{plot_par()}}{
Plot observation parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$plot_par(var, covs = NULL, n_post = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var}}{Name of covariate as a function of which the parameters
should be plotted}

\item{\code{covs}}{Optional data frame with a single row and one column
for each covariate, giving the values that should be used. If this is
not specified, the mean value is used for numeric variables, and the
first level for factor variables.}

\item{\code{n_post}}{Number of posterior draws to plot. Default: 0.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SDE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
