% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{signed_angle}
\alias{signed_angle}
\title{Compute signed angle in [-pi,pi] that rotates first vector into second vector}
\usage{
signed_angle(u, v)
}
\arguments{
\item{u}{matrice of bivariate vectors with 2 columns}

\item{v}{matrice of bivariate vectors with 2 columns}
}
\value{
vector of signed angles in [-pi,pi] that rotates first vector into second vector
# https://math.stackexchange.com/questions/529555/signed-angle-between-2-vectors
}
\description{
Compute signed angle in [-pi,pi] that rotates first vector into second vector
}
