% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothSDE-package.R
\docType{package}
\name{smoothSDE-package}
\alias{smoothSDE}
\alias{smoothSDE-package}
\title{smoothSDE: Varying-coefficient stochastic differential equations}
\description{
Provides functions to fit stochastic differential
    equations with time-varying parameters.
}
\details{
The package smoothSDE is based on the class \link{SDE}. Type ?SDE 
to find its documentation, or consult the package vignette for detailed 
examples of the smoothSDE workflow.
}
\keyword{internal}
