% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{split_term}
\alias{split_term}
\title{Split string with parenthesis}
\usage{
split_term(string, sep)
}
\arguments{
\item{string}{a string to split based on character sep}

\item{sep}{separation parameter}
}
\value{
list of string split based on sep, but ignoring sep that are within parenthesis
}
\description{
Split string with parenthesis
}
