% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{term_indices}
\alias{term_indices}
\title{Indices of coefficients for given model term}
\usage{
term_indices(names_fe, names_re, term)
}
\arguments{
\item{names_fe}{Names of fixed effect coefficients}

\item{names_re}{Names of random effect coefficients}

\item{term}{Name of term as character string, e.g. "time", 
or "s(time)"}
}
\value{
List with elements \code{fe} and \code{re}, which are vectors
of the indices in coeff_fe and coeff_re (respectively) corresponding
to the term.
}
\description{
Indices of coefficients for given model term
}
